+++ Krowi_AchievementFilter/Data/ExportedAchievements.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-07-26 13-12-10 ]] --
+-- [[ Exported at 2021-08-06 10-22-49 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --