+++ Krowi_AchievementFilter/Data/ExportedCalendarEvents.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-07-31 23-47-26 ]] --
+-- [[ Exported at 2021-08-06 14-59-22 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -7,10 +7,10 @@ local _, addon = ...;
 local objects = addon.Objects;
 local event = objects.Event;
 local data = addon.Data;
-data.ExportedEvents = {};
-local exportedEvents = data.ExportedEvents;
+data.ExportedCalendarEvents = {};
+local exportedCalendarEvents = data.ExportedCalendarEvents;
 
-function exportedEvents.Load(e)
+function exportedCalendarEvents.Load(e)
     for i, _ in next, e do
         e[i] = nil;
     end
@@ -29,22 +29,22 @@ function exportedEvents.Load(e)
     e[479] = event:New(479, 134481); -- Darkmoon Faire
 end
 
-function exportedEvents.LoadCategories(e, a)
+function exportedCalendarEvents.LoadCategories(e, a)
     if e[141] == nil or e[181] == nil or e[201] == nil or e[324] == nil or e[327] == nil or e[341] == nil or e[372] == nil or e[398] == nil or e[404] == nil or e[409] == nil or e[423] == nil or e[479] == nil then
         exportedEvents.Load(e);
     end
 
-    e[141].Category = a[273].Category; -- Winter Veil
-    e[181].Category = a[2676].Category; -- Noblegarden
-    e[201].Category = a[1791].Category; -- Children's Week
-    e[324].Category = a[972].Category; -- Hallow's End
-    e[327].Category = a[605].Category; -- Lunar Festival
-    e[341].Category = a[263].Category; -- Midsummer
-    e[372].Category = a[1184].Category; -- Brewfest
-    e[398].Category = a[3457].Category; -- Pirates' Day
-    e[404].Category = a[3579].Category; -- Pilgrim's Bounty
-    e[409].Category = a[3456].Category; -- Day of the Dead
-    e[423].Category = a[1701].Category; -- Love is in the Air
-    e[479].Category = a[6019].Category; -- Darkmoon Faire
+    e[141].Category = data.CategoriesEvents[1].Children[11]; -- Winter Veil
+    e[181].Category = data.CategoriesEvents[1].Children[3]; -- Noblegarden
+    e[201].Category = data.CategoriesEvents[1].Children[4]; -- Children's Week
+    e[324].Category = data.CategoriesEvents[1].Children[8]; -- Hallow's End
+    e[327].Category = data.CategoriesEvents[1].Children[1]; -- Lunar Festival
+    e[341].Category = data.CategoriesEvents[1].Children[5]; -- Midsummer
+    e[372].Category = data.CategoriesEvents[1].Children[7]; -- Brewfest
+    e[398].Category = data.CategoriesEvents[1].Children[6]; -- Pirates' Day
+    e[404].Category = data.CategoriesEvents[1].Children[10]; -- Pilgrim's Bounty
+    e[409].Category = data.CategoriesEvents[1].Children[9]; -- Day of the Dead
+    e[423].Category = data.CategoriesEvents[1].Children[2]; -- Love is in the Air
+    e[479].Category = data.CategoriesEvents[2]; -- Darkmoon Faire
 end